// backend/ResursPortal/frontend/src/features/credit-application/creditApplication.types.ts
export type CreditApplicationData = {
  orgNumber: string
  companyName: string
  authorizedSignatory: string

  equity: string
  totalCapital: string
  currentAssets: string
  shortTermLiabilities: string
  totalDebt: string
  operatingProfit: string
  netSales: string

  requestedAmount: string
  purpose: string
}